requirements.txt
requests>=2.31.0
httpx>=0.25.0
pandas>=2.0.0
pyarrow>=14.0.0
duckdb>=0.9.0
geopandas>=0.14.0
shapely>=2.0.0
pyyaml>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
click>=8.1.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
